Add daemon to TUID service with a local test, and fix some bugs.
This patch adds a daemon to the service that can be run to update
the frontier files to the latest known coverage revisions. It also
adds a test that can be ran locally to watch the daemon run. To make
the daemon possible, some other changes aalso needed to be made to
other portions of the service (like `get_tuids_from_files`).