Add stress test client and all required changes.
This patch adds a client that performs a stress test
on the service by querying for 1000 file chunks every
X seconds, at different revisions, where X is
decreasing over time. It stops when there are too
many items to process in the queue and only after
this happens more than once. tests/stress_client.py requires
that tuid/app.py and tests/sqs_consumer.py are running
for the stress test to be performed.